@@ -20,11 +20,8 @@ use App\Service\SportInfoProvider;
 
 class TeamType extends AbstractType
 {
-    private $sportInfo;
-
-    public function __construct(SportInfoProvider $sportInfo)
+    public function __construct(private readonly SportInfoProvider $sportInfo)
     {
-        $this->sportInfo = $sportInfo;
     }
 
     public function buildForm(FormBuilderInterface $builder, array $options)
