@@ -14,7 +14,7 @@ use App\Repository\DocumentRepository;
 class AdminController extends AbstractController
 {
 
-    public function __construct(private ManagerRegistry $doctrine) {}
+    public function __construct(private readonly ManagerRegistry $doctrine) {}
 
     /**
      * @Route(
