@@ -19,7 +19,7 @@ use Doctrine\Persistence\ManagerRegistry;
 class SearchController extends AbstractController
 {
 
-    public function __construct(private ManagerRegistry $doctrine) {}
+    public function __construct(private readonly ManagerRegistry $doctrine) {}
 
     /**
      * @Route(
