@@ -17,7 +17,7 @@ use Doctrine\Persistence\ManagerRegistry;
 class SeasonController extends AbstractController
 {
 
-    public function __construct(private ManagerRegistry $doctrine) {}
+    public function __construct(private readonly ManagerRegistry $doctrine) {}
 
     /**
      * @Route(
